+++ tests.py
@@ -4,7 +4,9 @@ import random
 from unittest import TestCase, main
 from contextlib import contextmanager
 import datetime
+
 import dbfread
+from dbfread.field_parser import FieldParser
 
 # http://docs.python.org/2/library/unittest.html                                
 
@@ -28,13 +30,80 @@ class MockField(object):
         self.type = type
         self.__dict__.update(kwargs)
 
-class TestFieldParsers(TestCase):
-    def test_M(self):
-        from dbfread.field_parser import FieldParser
-        parser = FieldParser('latin1')
+def make_field_parser(field_type):
+    parser = FieldParser('latin1')
+    field = MockField(field_type)
 
-        def parse(data):
-            return parser.parseM(MockField(), data)
+    def parse(data):
+        return parser.parse(field, data)
+
+    return parse
+
+class TestFieldParsers(TestCase):
+    def test_0(self):
+        parse = make_field_parser('0')
+        
+        assert parse(b'\x01') == 1
+        assert parse(b'\x00') == 0
+
+    def test_C(self):
+        parse = make_field_parser('C')
+        
+        assert type(parse(b'test')) == type(u'')
+
+    def test_D(self):
+        parse = make_field_parser('D')
+
+        assert parse(b'00000000') is None
+        assert parse(b'        ') is None
+
+        epoch = datetime.date(1970, 1, 1)
+        assert parse(b'19700101') == epoch
+
+        with raises(ValueError):
+            parse(b'NotIntgr')
+
+    def test_F(self):
+        parse = make_field_parser('F')
+
+        assert parse(b'') is None
+        assert parse(b' ') is None
+        
+        assert parse(b'0') == 0
+        assert parse(b'1') == 1
+        assert parse(b'-1') == -1
+        assert parse(b'3.14') == 3.14
+
+        with raises(ValueError):
+            parse(b'jsdf')
+
+    def test_I(self):
+        parse = make_field_parser('I')
+
+        # Little endian unsigned integer.
+        assert parse(b'\x00\x00\x00\x00') == 0
+        assert parse(b'\x01\x00\x00\x00') == 1
+        assert parse(b'\xff\xff\xff\xff') == -1
+
+    def test_L(self):
+        parse = make_field_parser('L')
+
+        for char in b'TtYy':
+            assert parse(char) is True
+
+        for char in b'FfNn':
+            assert parse(char) is False
+
+        for char in b'? ':
+            assert parse(char) is None
+
+        # Some invalid values.
+        for char in b'!0':
+            with raises(ValueError):
+                parse(char)
+
+    def test_M(self):
+        parse = make_field_parser('M')
 
         assert parse(b'\x01\x00\x00\x00') == 1
         assert parse(b'1') == 1
@@ -42,20 +111,24 @@ class TestFieldParsers(TestCase):
         with raises(ValueError):
             parse(b'NotInteger')
 
-    def test_D(self):
-        from dbfread.field_parser import FieldParser
-        parser = FieldParser('latin1')
+    def test_N(self):
+        parse = make_field_parser('N')
 
-        def parse(data):
-            return parser.parseD(MockField(), data)
-
-        assert parse(b'00000000') is None
-        assert parse(b'        ') is None
-
-        epoch = datetime.date(1970, 1, 1)
-        assert parse(b'19700101') == epoch
+        assert parse(b'') is None
+        assert parse(b' ') is None
+        assert parse(b'1') == 1
+        assert parse(b'-99') == -99
+        assert parse(b'3.14') == 3.14
         with raises(ValueError):
-            parse(b'NotIntgr')
+            parse(b'okasd')
+
+    def test_T(self):
+        parse = make_field_parser('T')
+
+        assert parse(b'') is None
+        assert parse(b' ') is None
+
+        # Todo: add more tests.
 
 if __name__ == '__main__':
     main()