+++ CHANGELOG.rst
@@ -3,6 +3,10 @@
 
 * removed circular dependency between table and deleted record iterator.
 
+* added ``dbversion`` attribute.
+
+* added example ``dbfinfo.py``.
+
 
 1.0.4 - 2014-07-27
 ------------------